# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Counterpoise API",
    "description": "The routes a native client uses. Money is integer cents. Shares and prices are integer micros. Dates are YYYY-MM-DD strings.",
    "version": "1.48.0",
    "x-api-contract": 1
  },
  "servers": [
    {
      "url": "/",
      "description": "The server this document was fetched from."
    }
  ],
  "paths": {
...
    "/api/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "The books you are a member of, with your role in each.",
        "operationId": "listBooks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookList"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Create a book.",
        "operationId": "createBook",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/books/{bookId}": {
      "put": {
        "tags": [
          "books"
        ],
        "summary": "Rename a book or change its settings.",
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "books"
        ],
        "summary": "Delete a book and everything in it.",
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/b/{bookId}/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Root accounts with balances and one level of children.",
        "operationId": "listAccounts",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Only accounts of this type: asset, liability, equity, income, or expense.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "description": "Send true to include inactive accounts.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "asOfDate",
            "in": "query",
            "description": "Balances as of this date as YYYY-MM-DD.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountNodeList"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create an account.",
        "operationId": "createAccount",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountNode"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountNode": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "name",
          "type",
          "subtype",
          "parentId",
          "isActive",
          "isFavorite",
          "isInvestmentCash",
          "icon",
          "createdAt",
          "updatedAt",
          "balance",
          "hasTransactions",
          "children"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "subtype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountSubtype"
              },
              {
                "type": "null"
              }
            ]
          },
          "parentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isInvestmentCash": {
            "type": "boolean"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "A single emoji, or null to inherit the parent's icon."
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "balance": {
            "$ref": "#/components/schemas/Cents"
          },
          "hasTransactions": {
            "type": "boolean"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountNode"
            }
          }
        }
      },
      "AccountNodeList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AccountNode"
        }
      },
...
      "AccountSubtype": {
        "type": "string",
        "enum": [
          "bank",
          "credit_card",
          "loan",
          "investment",
          "cash",
          "other"
        ]
      },
      "AccountType": {
        "type": "string",
        "enum": [
          "asset",
          "liability",
          "equity",
          "income",
          "expense"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A message safe to show to the user."
          }
        }
      },
...
      "Book": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "name",
          "upcomingDays",
          "typesafeReconciliationEnabled",
          "typesafeRevision",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "upcomingDays": {
            "type": "integer"
          },
          "typesafeReconciliationEnabled": {
            "type": "boolean"
          },
          "typesafeRevision": {
            "type": "integer"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "role": {
            "type": "string",
            "description": "Your role in this book: owner, editor or viewer. Present in the book list.",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ]
          }
        }
      },
      "BookList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Book"
        }
      },
...
      "Cents": {
        "type": "integer",
        "description": "Money in integer cents. Positive is a debit, negative is a credit."
      },
      "CreateAccountBody": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Account name, as it will appear in the chart of accounts.",
            "minLength": 1
          },
          "type": {
            "type": "string",
            "description": "Account type: asset, liability, equity, income, or expense. You cannot change this after creation.",
            "enum": [
              "asset",
              "liability",
              "equity",
              "income",
              "expense"
            ]
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bank",
                  "credit_card",
                  "loan",
                  "investment",
                  "cash",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "description": "Account subtype: bank, credit_card, loan, investment, cash, or other. Set to 'investment' to also create a paired cash sub-account automatically."
          },
          "parentId": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "description": "ID of the parent account, to nest this account under it. Omit for a top-level account. Use get_account_tree to find a parentId."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "A single emoji to show for this account. Only income and expense accounts show an icon. Send null to inherit the icon from the parent account instead of setting one directly."
          }
        }
      },
...
      "CreateBookBody": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The book's name, shown in the book switcher and page titles.",
            "minLength": 1
          }
        }
      },
...
      "Success": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "Timestamp": {
        "type": "string",
        "format": "date-time",
        "description": "A UTC instant as an ISO 8601 string with a Z suffix."
      },
...
      "UpdateBookBody": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The book's name, shown in the book switcher and page titles.",
            "minLength": 1
          },
          "upcomingDays": {
            "type": "integer",
            "description": "How many days ahead recurring transactions are projected on the dashboard (integer, 1-365). Omit to leave it unchanged.",
            "maximum": 365,
            "minimum": 1
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key: Authorization: Bearer cpk_..."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "counterpoise_session"
      }
    }
  }
}